% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{modeler}
\alias{modeler}
\title{Modeler}
\usage{
modeler(df, pERPs)
}
\arguments{
\item{df}{A dataframe containing a single record from the observed data.}

\item{pERPs}{The estimated pERPs from the pERP-RED algorithm.}
}
\value{
The linear model from regressing the observed record on the estimated pERPs.
}
\description{
This function will regress the individual record on the estimated pERPs, we will need to demean the record before regression.
}
