% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pERP_difference.R
\name{pERP_difference}
\alias{pERP_difference}
\title{Calculate t-tests for pERPs}
\usage{
pERP_difference(
  scores,
  electrode,
  task1,
  group1 = "Overall",
  task2 = task1,
  group2 = group1,
  double_difference = FALSE
)
}
\arguments{
\item{scores}{The dataframe with all of the individual weights.}

\item{electrode}{A string of the electrode for testing.}

\item{task1}{A string of the first task for comparison.}

\item{group1}{A string of the first treatment/diagnosis group to test. If "Overall", all of the individuals will be averaged.}

\item{task2}{A string of the second task for comparison, set to the same value as task1 by default.}

\item{group2}{A string of the second treatment/diagnosis group to test. If "Overall", all of the individuals will be averaged.}

\item{double_difference}{A logical to contrast task1 - task2. Use in cases such as correct - incorrect.}
}
\value{
A dataframe of t-test results
}
\description{
This function will take in two comparison groups.
}
