% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{pERP_scorer}
\alias{pERP_scorer}
\title{pERP_scorer}
\usage{
pERP_scorer(df, pERPs)
}
\arguments{
\item{df}{A dataframe containing all of the observed ERP records.}

\item{pERPs}{The estimated pERPs from the pERP-RED algorithm.}
}
\value{
The dataframe containing each individual score for each of the records.
}
\description{
This function calculates all of the individual score values.
}
