% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_pERP.R
\name{R2_pERP}
\alias{R2_pERP}
\title{R2_pERP}
\usage{
R2_pERP(true_pERPs, pERPs)
}
\arguments{
\item{true_pERPs}{The dataframe containing the true pERPs.}

\item{pERPs}{The dataframe containing the estimated pERPs.}
}
\value{
a dataframe containing the number of pERPs and the associated R2_pERP value
}
\description{
This function will regress the true pERPs on the estimated pERPs.
}
