% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_test.R
\name{R2_test}
\alias{R2_test}
\title{R2_test}
\usage{
R2_test(df, pERPs)
}
\arguments{
\item{df}{The observed data, typically the test set.}

\item{pERPs}{The estimated pERPs from the pERP-RED algorithm.}
}
\value{
a dataframe containing the number of pERPs and the associated R2_test value
}
\description{
This function will regress one record from the observed data on the estimated pERPs at a time and calculate a global R2 value.
}
