% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pERPs.R
\name{shiny_pERPs}
\alias{shiny_pERPs}
\title{Shiny App pERP Explorer}
\usage{
shiny_pERPs(data, groups, pERPs, scores)
}
\arguments{
\item{data}{A dataframe containing all of the ERP data used to calculate the pERPs. One column for Task, Subject, and Time, and the rest are electrodes with the ERP data in the rows.}

\item{groups}{A dataframe containing Subject, which should have the same form as the Subject column in the data, and Group, a group membership indicator.}

\item{pERPs}{A dataframe containing the pERPs. This is the resulting dataframe from running `pERPred`.}

\item{scores}{A dataframe containing each of the individual scores from regressing the observed data on the pERPs. This is the resulting dataframe from running `pERP_scorer`.}
}
\value{
shiny application
}
\description{
In our paper, we often referenced a Shiny app in order to explore all of the results from the pERP-space analysis. Now you can, too! This function will take all of your data and results and launch a shiny application on your machine.
}
