% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_test.R
\name{sum_of_squares_calc}
\alias{sum_of_squares_calc}
\title{sum_of_squares_calc}
\usage{
sum_of_squares_calc(df, pERPs)
}
\arguments{
\item{df}{The observed data, typically the test set.}

\item{pERPs}{The estimated pERPs from the pERP-RED algorithm.}
}
\value{
a dataframe containing the number of pERPs and the associated R2_test value
}
\description{
This function will regress one record from the observed data on the estimated pERPs at a time and calculate the variance of the residuals and the variance of the observed signal to be used in the global R2 calculation.
}
