% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pERPred.R
\name{pERPred}
\alias{pERPred}
\title{Principle ERP Reduction}
\usage{
pERPred(df, num_pERPs = 20, percent_variation = 80)
}
\arguments{
\item{df}{The dataframe with a column for Task, Subject, and Time. The remaining columns are the Electrodes.}

\item{num_pERPs}{The number of pERPs to estimate.}

\item{percent_variation}{The percent variation to use in the PCA steps.}
}
\value{
pERPs The prinicple ERPs (pERPs) are the bases functions estimated by the pERP-RED algorithm.
}
\description{
This function will contain the estimation algorithm of the principle ERPs (pERPs).
}
